refactor(api): centralize all quota constants into quota.constants.ts
