--- apps/api/src/services/bucket.constants.ts
@@ -1,2 +1,0 @@
-/** Maximum number of non-archived privacy buckets per system. */
-export const MAX_BUCKETS_PER_SYSTEM = 50;
