+++ apps/api/src/services/bucket.service.ts
@@ -17,13 +17,13 @@ import { parseQuery } from "../lib/query-parse.js";
 import { withTenantRead, withTenantTransaction } from "../lib/rls-context.js";
 import { assertSystemOwnership } from "../lib/system-ownership.js";
 import { tenantCtx } from "../lib/tenant-context.js";
+import { MAX_BUCKETS_PER_SYSTEM } from "../quota.constants.js";
 import {
   DEFAULT_PAGE_LIMIT,
   MAX_ENCRYPTED_DATA_BYTES,
   MAX_PAGE_LIMIT,
 } from "../service.constants.js";
 
-import { MAX_BUCKETS_PER_SYSTEM } from "./bucket.constants.js";
 import { dispatchWebhookEvent } from "./webhook-dispatcher.js";
 
 import type { AuditWriter } from "../lib/audit-writer.js";
