+++ apps/api/src/services/friend-code.constants.ts
@@ -1,6 +1,3 @@
-/** Maximum number of active (non-archived) friend codes per account. */
-export const MAX_FRIEND_CODES_PER_ACCOUNT = 10;
-
 /** Number of random bytes used to generate the friend code string. */
 export const FRIEND_CODE_BYTES = 8;
