+++ apps/api/src/services/innerworld-entity.service.ts
@@ -12,6 +12,7 @@ import { buildPaginatedResult } from "../lib/pagination.js";
 import { withTenantRead, withTenantTransaction } from "../lib/rls-context.js";
 import { assertSystemOwnership } from "../lib/system-ownership.js";
 import { tenantCtx } from "../lib/tenant-context.js";
+import { MAX_INNERWORLD_ENTITIES_PER_SYSTEM } from "../quota.constants.js";
 import {
   DEFAULT_PAGE_LIMIT,
   MAX_ENCRYPTED_DATA_BYTES,
@@ -29,8 +30,6 @@ import type {
   UnixMillis,
 } from "@pluralscape/types";
 import type { PostgresJsDatabase } from "drizzle-orm/postgres-js";
-
-const MAX_INNERWORLD_ENTITIES_PER_SYSTEM = 500;
 
 // ── Types ───────────────────────────────────────────────────────────
