+++ apps/api/src/services/innerworld-region.service.ts
@@ -11,6 +11,7 @@ import { buildPaginatedResult } from "../lib/pagination.js";
 import { withTenantRead, withTenantTransaction } from "../lib/rls-context.js";
 import { assertSystemOwnership } from "../lib/system-ownership.js";
 import { tenantCtx } from "../lib/tenant-context.js";
+import { MAX_INNERWORLD_REGIONS_PER_SYSTEM } from "../quota.constants.js";
 import {
   DEFAULT_PAGE_LIMIT,
   MAX_ENCRYPTED_DATA_BYTES,
@@ -27,11 +28,6 @@ import type {
   UnixMillis,
 } from "@pluralscape/types";
 import type { PostgresJsDatabase } from "drizzle-orm/postgres-js";
-
-// ── Constants ───────────────────────────────────────────────────────
-
-/** Maximum number of active innerworld regions allowed per system. */
-const MAX_INNERWORLD_REGIONS_PER_SYSTEM = 100;
 
 // ── Types ───────────────────────────────────────────────────────────
